move kind_to_link_type to netlink-utils.hpp
